Retry upon Azure GPT-4 model overloaded requests error
