+++ app/page.tsx
@@ -1,0 +1,8 @@
+import Image from 'next/image'
+
+export default function Home() {
+  return (
+    <main>
+    </main>
+  )
+}
