+++ package.json
@@ -7,25 +7,16 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "lint:fix": "eslint . --fix",
-    "lint:strict": "eslint --max-warnings=0 .",
-    "format": "prettier --write '**/*.{ts,tsx,js,html,css,md}'",
-    "format:check": "prettier --check '**/*.{ts,tsx,js,html,css,md}'"
+    "format": "prettier --write '**/*.{ts,tsx,js,html,css,md}'"
   },
   "dependencies": {
-    "axios": "^1.4.0",
-    "next": "13.4.3",
+    "@types/node": "20.4.4",
+    "@types/react": "18.2.15",
+    "@types/react-dom": "18.2.7",
+    "next": "13.4.12",
+    "prettier": "^3.0.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "winston": "^3.8.2"
-  },
-  "devDependencies": {
-    "@types/node": "20.2.1",
-    "@types/react": "18.2.6",
-    "@types/react-dom": "18.2.4",
-    "eslint": "8.40.0",
-    "eslint-config-next": "13.4.3",
-    "prettier": "^2.8.8",
-    "typescript": "5.0.4"
+    "typescript": "5.1.6"
   }
 }
