--- src/pages/api/chat/interface.ts
@@ -1,10 +1,0 @@
-export enum ChatRole {
-  Assistant = 'assistant',
-  User = 'user',
-  System = 'system'
-}
-
-export interface ChatMessage {
-  content: string
-  role: ChatRole
-}
